Split `NodeType` from `python.dom.node.Node`
