+++ python/dom/node/Node.py
@@ -1,4 +1,3 @@
-import enum
 import warnings
 from typing import AnyStr, Optional
 
@@ -8,25 +7,7 @@ from python.dom.NotImplemented.NamedNodeMap import NamedNodeMap
 from python.dom.NodeList import NodeList
 from python.dom.type_checking import AnyNode
 from python.dom.node.Document import Document
-
-
-class NodeType(enum.IntEnum):
-    """Definition group `NodeType`
-
-    An integer indicating which type of node this is.
-    """
-    ELEMENT_NODE = 1  # The node is a `Element`.
-    ATTRIBUTE_NODE = 2  # The node is an `Attr`.
-    TEXT_NODE = 3  # The node is a `Text` node.
-    CDATA_SECTION_NODE = 4  # The node is a `CDATASection`.
-    ENTITY_REFERENCE_NODE = 5  # The node is an `EntityReference`.
-    ENTITY_NODE = 6  # The node is an `Entity`.
-    PROCESSING_INSTRUCTION_NODE = 7  # The node is a `ProcessingInstruction`.
-    COMMENT_NODE = 8  # The node is a `Comment`.
-    DOCUMENT_NODE = 9  # The node is a `Document`.
-    DOCUMENT_TYPE_NODE = 10  # The node is a `DocumentType`.
-    DOCUMENT_FRAGMENT_NODE = 11  # The node is a `DocumentFragment`.
-    NOTATION_NODE = 12  # The node is a `Notation`.
+from python.dom.node.NodeType import NodeType
 
 
 class Node:
@@ -36,18 +17,6 @@ class Node:
 
     The attributes `node_name`, `node_value` and attributes are included as a mechanism to get at node information without casting down to the specific derived interface. In cases where there is no obvious mapping of these attributes for a specific `node_type` (e.g., `node_value` for an Element or `attributes` for a Comment), this returns `None`. Note that the specialized interfaces may contain additional and more convenient mechanisms to get and set the relevant information.
     """
-    ATTRIBUTE_NODE = NodeType.ATTRIBUTE_NODE
-    TEXT_NODE = NodeType.TEXT_NODE
-    ELEMENT_NODE = NodeType.ELEMENT_NODE
-    CDATA_SECTION_NODE = NodeType.CDATA_SECTION_NODE
-    ENTITY_REFERENCE_NODE = NodeType.ENTITY_REFERENCE_NODE
-    ENTITY_NODE = NodeType.ENTITY_NODE
-    PROCESSING_INSTRUCTION_NODE = NodeType.PROCESSING_INSTRUCTION_NODE
-    COMMENT_NODE = NodeType.COMMENT_NODE
-    DOCUMENT_NODE = NodeType.DOCUMENT_NODE
-    DOCUMENT_TYPE_NODE = NodeType.DOCUMENT_TYPE_NODE
-    DOCUMENT_FRAGMENT_NODE = NodeType.DOCUMENT_FRAGMENT_NODE
-    NOTATION_NODE = NodeType.NOTATION_NODE
 
     def __init__(self,
                  owner_document: Document,
@@ -111,9 +80,9 @@ class Node:
         if node_type is None:
             node_type = self.node_type
         if parent_node is not None:
-            if self.node_type in [Node.DOCUMENT_NODE,
-                                  Node.DOCUMENT_FRAGMENT_NODE,
-                                  Node.ATTRIBUTE_NODE]:
+            if self.node_type in [NodeType.DOCUMENT_NODE,
+                                  NodeType.DOCUMENT_FRAGMENT_NODE,
+                                  NodeType.ATTRIBUTE_NODE]:
                 warnings.warn('Document, DocumentFragment, '
                               'and Attr may not have a parent.')
         self._parent_node = parent_node
